0.2.4
Ивенты на использование абилки + конец абилки + конец кд
Новая система здоровья и урона, с сопротивлением урону
Реворк системы стрельбы ( теперь рейкасты)

+++ Assets/Scripts/ScriptableObjects/WeaponConfig.cs
@@ -5,7 +5,7 @@ using UnityEngine.UIElements;
 public class WeaponConfig : ScriptableObject, IInfoBlockUI
 {
     public float damage;
-    public int fireRate;
+    public float fireRate;
     public float range;
     public WeaponTypes weaponType;
     public float typeStat;
